Make the vite compat test more comprehensive
